[refactor] Create files for comparison checker in pylint.checker.base
